Revert "Implemented BaseGame"
This reverts commit 9523a748f4a1a8915c5d93a1158fe2e69c22b538.
